refactor: reservation.entity 패키지를 reservation.domain으로 변경
