+++ src/main/java/kr/hhplus/be/server/reservation/domain/Reservation.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.reservation.entity;
+package kr.hhplus.be.server.reservation.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
