+++ src/main/java/kr/hhplus/be/server/reservation/domain/SeatStatus.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.reservation.entity;
+package kr.hhplus.be.server.reservation.domain;
 
 public enum SeatStatus {
   AVAILABLE, HOLD, RESERVED
