+++ src/test/java/kr/hhplus/be/server/reservation/ReservationControllerTest.java
@@ -16,8 +16,8 @@ import kr.hhplus.be.server.reservation.adapter.in.web.ReservationController;
 import kr.hhplus.be.server.reservation.adapter.in.web.dto.request.ReservationRequest;
 import kr.hhplus.be.server.reservation.application.port.in.ReservationResult;
 import kr.hhplus.be.server.reservation.application.port.in.ReserveSeatUseCase;
-import kr.hhplus.be.server.reservation.entity.Reservation;
-import kr.hhplus.be.server.reservation.entity.Seat;
+import kr.hhplus.be.server.reservation.domain.Reservation;
+import kr.hhplus.be.server.reservation.domain.Seat;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
