+++ src/test/java/kr/hhplus/be/server/reservation/ReservationServiceTest.java
@@ -9,9 +9,9 @@ import kr.hhplus.be.server.reservation.application.port.out.LoadUserPort;
 import kr.hhplus.be.server.reservation.application.port.out.SaveReservationPort;
 import kr.hhplus.be.server.reservation.application.port.out.UpdateSeatPort;
 import kr.hhplus.be.server.reservation.application.service.ReservationService;
-import kr.hhplus.be.server.reservation.entity.Reservation;
-import kr.hhplus.be.server.reservation.entity.Seat;
-import kr.hhplus.be.server.reservation.entity.SeatStatus;
+import kr.hhplus.be.server.reservation.domain.Reservation;
+import kr.hhplus.be.server.reservation.domain.Seat;
+import kr.hhplus.be.server.reservation.domain.SeatStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
